requirements.txt
wxPython==4.0.0b2
Pillow==4.3.0